Accept tif input in 2D DNNs

+++ segment/_2D_DNN/ExecuteInference.py
@@ -38,20 +38,22 @@ class ExecuteInference(MiscellaneousSegment):
 
 
         input_files = glob.glob(os.path.join(params['Image Folder'], "*.jpg"))
+        input_png = glob.glob(os.path.join(params['Image Folder'], "*.png"))
+        input_tif = glob.glob(os.path.join(params['Image Folder'], "*.tif"))
+        input_files.extend(input_png)
+        input_files.extend(input_tif)
         if len(input_files) == 0:
-            input_files = glob.glob(os.path.join(params['Image Folder'], "*.png"))
-            if len(input_files) == 0:
-                print('No images in the Image Folder.')
-                return
+            print('No images in the Image Folder.')
+            return
 
         im = cv2.imread(input_files[0], cv2.IMREAD_UNCHANGED)
         print('Target file to check color type : ', input_files[0])
         print('Image dimensions                : ', im.shape)
         print('Image filetype                  : ', im.dtype)
-        image_width  = im.shape[0]
-        image_height = im.shape[1]
+        image_width  = im.shape[1]
+        image_height = im.shape[0]
 
-        if not (im.dtype == "uint8" and len(im.shape) == 3) :
+        if not (im.dtype == "uint8" and len(im.shape) == 3 and  input_tif == []) :
             tmpdir = os.path.join(datadir, "tmp", "DNN_test_images")
             if os.path.exists(tmpdir) :
                 shutil.rmtree(tmpdir)
@@ -59,6 +61,7 @@ class ExecuteInference(MiscellaneousSegment):
             for input_file in input_files:
                 im_col = cv2.imread(input_file)
                 filename = os.path.basename(input_file)
+                filename = filename.replace('.tif', '.png')
                 converted_input_file = os.path.join( tmpdir, filename )
                 cv2.imwrite(converted_input_file, im_col)
             params['Image Folder'] = tmpdir
